+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v1/EntityGraphRetriever.java
@@ -286,7 +286,7 @@ public final class EntityGraphRetriever {
         entity.setGuid(GraphHelper.getGuid(entityVertex));
         entity.setTypeName(GraphHelper.getTypeName(entityVertex));
         entity.setStatus(GraphHelper.getStatus(entityVertex));
-        entity.setVersion(GraphHelper.getVersion(entityVertex).longValue());
+        entity.setVersion(GraphHelper.getVersion(entityVertex));
 
         entity.setCreatedBy(GraphHelper.getCreatedByAsString(entityVertex));
         entity.setUpdatedBy(GraphHelper.getModifiedByAsString(entityVertex));
@@ -776,7 +776,7 @@ public final class EntityGraphRetriever {
         relationship.setCreateTime(new Date(GraphHelper.getCreatedTime(edge)));
         relationship.setUpdateTime(new Date(GraphHelper.getModifiedTime(edge)));
 
-        Integer version = GraphHelper.getVersion(edge);
+        Integer version = GraphHelper.getVersion(edge).intValue();
 
         if (version == null) {
             version = Integer.valueOf(1);
